We're moving to a Python 3.8+ minimum requirement
We're in the process to update `base`/`base-web` Docker images to use
a newer version of Ubuntu (`20.04`) which comes with Python 3.8 and
we're updating all repositories to suggest using 3.8 minumum.

Although this application does **NOT** uses `base-web` as a base image
(in fact it's already using Python 3.9) it makes sense to update this
README for consistency.

Also updated `clean-ecr` GH Workflow to use newer Ubuntu 20.04 container.

Related to ticket: https://dsdmoj.atlassian.net/browse/MTP-1841